pathish

+++ EECS494-Assignment4/Assets/PathingNode.cs
@@ -5,7 +5,7 @@ public class PathingNode {
 	public bool pathable;
 	public int x;
 	public int z;
-    public int heu;
+    public double heu;
 	public PathingNode bestNode;
 	public int dir;
